align implementation with FIP
The FIP breaks the logic down into:

1. Codec specific logic.
2. General "scan" logic.

I'm doing the same here as it makes it easier to reason about the code
and will help us in the future if we decide to support additional
codecs.
